REL-861575: Finalized the teardown/workspace delete issue with a TryDeleteWorkspace method.

+++ Source/kCura.IntegrationPoints.FunctionalTests/Functional/Helpers/RelativityFacadeExtensions.cs
@@ -41,21 +41,25 @@ namespace Relativity.IntegrationPoints.Tests.Functional.Helpers
             return instance.Resolve<IWorkspaceService>().Get(workspaceArtifactID);
         }
 
-        public static void DeleteWorkspace(this IRelativityFacade instance, Workspace workspace)
+        public static void TryDeleteWorkspace(this IRelativityFacade instance, Workspace workspace)
         {
             IWorkspaceService workspaceService = instance.Resolve<IWorkspaceService>();
+            try
+            {
+                if (!workspaceService.Exists(workspace.ArtifactID))
+                {
+                    return;
+                }
 
-            workspaceService.Delete(workspace.ArtifactID);
+                workspaceService.Delete(workspace.ArtifactID);
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine($"Functional test failed to delete workspace {workspace.ArtifactID}. Error: {e}");
+            }
         }
 
-        public static bool WorkspaceExists(this IRelativityFacade instance, Workspace workspace)
-        {
-	        IWorkspaceService workspaceService = instance.Resolve<IWorkspaceService>();
-
-	        return workspaceService.Exists(workspace.ArtifactID);
-        }
-
-		public static void RequireAgent(this IRelativityFacade instance, string agentTypeName, int runInterval)
+        public static void RequireAgent(this IRelativityFacade instance, string agentTypeName, int runInterval)
         {
             IAgentService agentService = instance.Resolve<IAgentService>();
             agentService.Require(new Agent
